arm64/mm: Enable soft-dirty PTE tracking
Alias PTE_PROT_NONE to PTE_SOFT_DIRTY to track the soft-dirty status of a
PTE. PTE_PROT_NONE is not interpreted by the hardware, and at the moment
it's meaningful for the kernel only when PTE_VALID is zero. On the other
hand, soft-dirty status is only meaningful for valid PTEs. Thus, it should
be possible to use it for both cases and achieve parity with x86, without
needing an additional PTE bit for soft-dirty tracking.

As discussed in the linked LKML thread of a previous attempt to enable
soft-dirty tracking for arm64, the software PTE_DIRTY bit is needed even
when using HW_AFDBM, to track the real dirty status of the PTE (e.g. when
write-protecting a dirty PTE without calling mkclean()), and by co-opting
it for soft-dirty tracking leads to an issue with erroneously resetting
soft-dirty status when using MADV_FREE.

The swap entry part is the same as in the original patch of the earlier
attempt, since the swap codepaths require !pte_present(), which means
PTE_PROT_NONE must be zero in the swap PTE, hence we need an extra bit for
tracking the soft-dirty status of the swap entry, but this is less
problematic since the swap PTE is invalid.

Link: https://lore.kernel.org/all/1512029649-61312-1-git-send-email-bin.lu@arm.com/
Signed-off-by: Stratos Psomadakis <psomas@cslab.ece.ntua.gr>